+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 -   Harlequin now returns the result of multiple select queries to different tabs in the Results Viewer. To run multiple queries, type them into the Query Editor (separated by semicolons), then press <kbd>ctrl+a</kbd> to select all, and then <kbd>ctrl+enter</kbd> to run the selection ([#34](https://github.com/tconbeer/harlequin/issues/34)).
+-   If there are multiple results tabs, you can switch between them with <kbd>j</kbd> and <kbd>k</kbd>.
 ## [0.0.24] - 2023-08-04
 
 -   Adds a new CLI option, `--extension` or `-e`, which will install and load a named DuckDB extension.